# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "administrator",
                "user",
                "auditor"
              ]
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/devices": {
      "get": {
        "operationId": "DevicesController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "collaborator",
                "infrastructure"
              ]
            }
          },
          {
            "name": "osName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hostname",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/inventory": {
      "post": {
        "operationId": "IngestionController_ingestInventory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryIngestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "node-api-key": []
          }
        ],
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/access-events": {
      "post": {
        "operationId": "IngestionController_ingestAccessEvent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessEventIngestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "node-api-key": []
          }
        ],
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/alert-rules": {
      "get": {
        "operationId": "AlertRulesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "alert-rules"
        ]
      },
      "post": {
        "operationId": "AlertRulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "alert-rules"
        ]
      }
    },
    "/api/v1/alert-rules/{id}": {
      "patch": {
        "operationId": "AlertRulesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "alert-rules"
        ]
      }
    },
    "/api/v1/alerts": {
      "get": {
        "operationId": "AlertsController_findAll",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "resource_change",
                "off_hours_access"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "reviewed"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/v1/alerts/{id}": {
      "patch": {
        "operationId": "AlertsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/v1/stats/devices": {
      "get": {
        "operationId": "StatsController_getDeviceStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "stats"
        ]
      }
    },
    "/api/v1/stats/os": {
      "get": {
        "operationId": "StatsController_getOsStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "stats"
        ]
      }
    },
    "/api/v1/reports/export": {
      "get": {
        "operationId": "ReportsController_export",
        "parameters": [
          {
            "name": "reportType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "devices",
                "alerts"
              ]
            }
          },
          {
            "name": "format",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "csv",
                "pdf"
              ]
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "collaborator",
                "infrastructure"
              ]
            }
          },
          {
            "name": "osName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hostname",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "alertType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "resource_change",
                "off_hours_access"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "reviewed"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "reports"
        ]
      }
    }
  },
  "info": {
    "title": "IT-MAS API",
    "description": "Management and Audit System — inventory ingestion, alerting, and portal query API.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "node-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Node-Api-Key"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "administrator",
              "user",
              "auditor"
            ]
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "administrator",
              "user",
              "auditor"
            ]
          },
          "active": {
            "type": "boolean"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        }
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "OsDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "version": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name",
          "version"
        ]
      },
      "CpuDto": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string",
            "minLength": 1
          },
          "cores": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "model",
          "cores"
        ]
      },
      "RamDto": {
        "type": "object",
        "properties": {
          "totalGB": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "totalGB"
        ]
      },
      "DiskDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "sizeGB": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "name",
          "sizeGB"
        ]
      },
      "InventoryIngestDto": {
        "type": "object",
        "properties": {
          "hostname": {
            "type": "string",
            "minLength": 1
          },
          "category": {
            "type": "string",
            "enum": [
              "collaborator",
              "infrastructure"
            ]
          },
          "os": {
            "$ref": "#/components/schemas/OsDto"
          },
          "cpu": {
            "$ref": "#/components/schemas/CpuDto"
          },
          "ram": {
            "$ref": "#/components/schemas/RamDto"
          },
          "disks": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiskDto"
            }
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": [
          "hostname",
          "category",
          "os",
          "cpu",
          "ram",
          "disks",
          "timestamp"
        ]
      },
      "AccessEventIngestDto": {
        "type": "object",
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "os",
              "database"
            ]
          },
          "user": {
            "type": "string",
            "minLength": 1
          },
          "timestamp": {
            "type": "string"
          },
          "action": {
            "type": "string",
            "enum": [
              "login",
              "logout"
            ]
          }
        },
        "required": [
          "level",
          "user",
          "timestamp",
          "action"
        ]
      },
      "HabitualHoursDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          }
        },
        "required": [
          "from",
          "to"
        ]
      },
      "AlertRuleConfigDto": {
        "type": "object",
        "properties": {
          "resources": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": [
                "cpu",
                "ram",
                "disks"
              ]
            }
          },
          "habitualHours": {
            "$ref": "#/components/schemas/HabitualHoursDto"
          }
        }
      },
      "CreateAlertRuleDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "resource_change",
              "off_hours_access"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "config": {
            "$ref": "#/components/schemas/AlertRuleConfigDto"
          }
        },
        "required": [
          "type",
          "config"
        ]
      },
      "UpdateAlertRuleDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "config": {
            "$ref": "#/components/schemas/AlertRuleConfigDto"
          }
        }
      },
      "UpdateAlertDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "reviewed"
            ]
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}